# backend/turismo-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Turismo Necochea",
        "description": "Documentación de la API para la aplicación de Turismo de Necochea",
        "contact": {
            "email": "tudai@alumnos.exa.unicen.edu.ar"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/tipos": {
            "get": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Obtener todos los tipos genéricos",
                "operationId": "ed6150ea38454e1926202e6ae515253d",
                "responses": {
                    "200": {
                        "description": "Lista de tipos genéricos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Crear un nuevo tipo genérico",
                "operationId": "66d9955b9219dc3770fedc979ea05af8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo genérico creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tipo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/tipos/{id}": {
            "get": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Obtener un tipo genérico por ID",
                "operationId": "2f7bbf172dc04c0ba03cf815198fd708",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo genérico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del tipo genérico",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tipo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo genérico no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Actualizar un tipo genérico por ID",
                "operationId": "f188186cef82f62d673a9633beba996c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo genérico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo genérico actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tipo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo genérico no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Eliminar un tipo genérico por ID",
                "operationId": "378a94fdd0fc70f233c83e387a0721b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo genérico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo genérico eliminado correctamente"
                    },
                    "404": {
                        "description": "Tipo genérico no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tipo": {
                "title": "Tipo",
                "description": "Modelo de un tipo genérico",
                "properties": {
                    "id": {
                        "description": "ID único del tipo",
                        "type": "integer",
                        "format": "int64"
                    },
                    "tipo": {
                        "description": "Nombre del tipo (ej. 'Aventura', 'Cultural')",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TipoRequest": {
                "title": "TipoRequest",
                "description": "Esquema para la creación de un tipo genérico",
                "required": [
                    "tipo"
                ],
                "properties": {
                    "tipo": {
                        "type": "string",
                        "example": "Aventura"
                    }
                },
                "type": "object"
            },
            "TipoUpdateRequest": {
                "title": "TipoUpdateRequest",
                "description": "Esquema para la actualización de un tipo genérico",
                "properties": {
                    "tipo": {
                        "type": "string",
                        "example": "Cultural"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}